refactor(web/users): update user management
