+++ apps/web/src/components/users/userHandling.tsx
@@ -2,25 +2,24 @@ import { Plus } from 'lucide-react';
 import { useEffect, useState } from 'react';
 import { toast } from 'sonner';
 
-import { SuperButton } from '../common/button';
-import { DataTable } from '../common/data-table/dataTable';
-
-import { UserDialog } from './UserDialog';
 import { columns } from './usersColumns';
 
+import { SuperButton } from '@/components/common/button';
+import { DataTable } from '@/components/common/data-table/dataTable';
+import { FormDialog } from '@/components/common/form/FormDialog';
 import {
   manageUserInputs,
   adminUserRole,
   supervisorUserRole,
 } from '@/lib/input';
 import { useUser } from '@/providers/UserProvider';
-import { getUsers, deleteUser, createUser } from '@/services/user.service';
+import { getUsers, createUser } from '@/services/user.service';
 
 export default function UserHandling() {
   const [data, setData] = useState<User[]>([]);
   const { user: currentUser } = useUser();
 
-  const [_loading, setLoading] = useState(false);
+  const [loading, setLoading] = useState(false);
 
   const roleInput =
     currentUser?.role === 'ADMIN' ? adminUserRole : supervisorUserRole;
@@ -31,7 +30,6 @@ export default function UserHandling() {
     try {
       setLoading(true);
       const users = await getUsers();
-      console.log(users);
       setData(users);
     } catch {
       toast.error('Erreur lors du chargement des utilisateurs');
@@ -40,19 +38,23 @@ export default function UserHandling() {
     }
   };
 
+  const handleCreate = async (data: CreateUserInput) => {
+    try {
+      setLoading(true);
+      await createUser(data);
+      toast.success('Utilisateur créé');
+      fetchUsers();
+    } catch (error: any) {
+      toast.error(error?.message || 'Utilisateur non créé');
+      return false;
+    } finally {
+      setLoading(false);
+    }
+  };
+
   useEffect(() => {
     fetchUsers();
   }, []);
-
-  const handleDelete = async (id: string) => {
-    try {
-      await deleteUser(id);
-      toast.success('Utilisateur supprimé');
-      fetchUsers();
-    } catch {
-      toast.error('Erreur lors de la suppression');
-    }
-  };
 
   return (
     <div className="container mx-auto h-full">
@@ -63,7 +65,7 @@ export default function UserHandling() {
         </h1>
       </div>
 
-      <UserDialog
+      <FormDialog
         trigger={
           <SuperButton
             voiceText={"Le paramètre avancé, c'est pour les grands."}
@@ -76,21 +78,13 @@ export default function UserHandling() {
         title="Créer un utilisateur"
         description="Ajouter un nouvel utilisateur"
         inputs={userInputs}
-        loading={false}
-        onsubmit={async data => {
-          console.log(data);
-          await createUser(data);
-          toast.success('Utilisateur créé');
-          fetchUsers();
-        }}
+        loading={loading}
+        onsubmit={handleCreate}
       />
 
       {/* TABLE */}
       <div>
-        <DataTable
-          columns={columns(handleDelete, fetchUsers, currentUser)}
-          data={data}
-        />
+        <DataTable columns={columns(fetchUsers, currentUser)} data={data} />
       </div>
     </div>
   );
