+++ apps/web/src/components/users/usersColumns.tsx
@@ -1,20 +1,10 @@
 import { ColumnDef } from '@tanstack/react-table';
-import { Pencil, Trash2 } from 'lucide-react';
-import { toast } from 'sonner';
 
-import { UserDeleteDialog } from './UserDeleteDialog';
-import { UserDialog } from './UserDialog';
+import { UserActions } from './UserActions';
 
 import { Checkbox } from '@/components/ui/checkbox';
-import {
-  manageUserInputs,
-  adminUserRole,
-  supervisorUserRole,
-} from '@/lib/input';
-import { updateUser } from '@/services/user.service';
 
 export const columns = (
-  onDelete: (id: string) => void,
   refresh: () => void,
   currentUser: UserSession | null
 ): ColumnDef<User>[] => [
@@ -92,57 +82,8 @@ export const columns = (
     id: 'actions',
     cell: ({ row }) => {
       const user = row.original;
-      const isSelf = currentUser!.id === user.id;
-      const roleInput =
-        currentUser?.role === 'ADMIN' ? adminUserRole : supervisorUserRole;
-
-      const userInputs = [...manageUserInputs, roleInput];
-
-      if (isSelf) return null;
-
       return (
-        <div className="flex items-center gap-2">
-          <UserDialog
-            trigger={
-              <button
-                className="p-2 rounded-full hover:bg-mauve/30 bg-mauve/10 text-mauve transition-colors"
-                title="Modifier"
-              >
-                <Pencil className="h-4 w-4" />
-              </button>
-            }
-            title="Modifier utilisateur"
-            description="Modifier les informations de l'utilisateur"
-            inputs={userInputs}
-            initialValues={{
-              email: user.email,
-              first_name: user.first_name,
-              last_name: user.last_name,
-              pseudo: user.profile?.pseudo,
-              role: user.role,
-            }}
-            loading={false}
-            onsubmit={async data => {
-              await updateUser(user.id, data);
-              toast.success('Utilisateur mis à jour');
-              refresh();
-            }}
-          />
-
-          <UserDeleteDialog
-            trigger={
-              <button
-                className="p-2 rounded-full  hover:bg-red-500/30 bg-red-500/10 text-red-500 transition-colors"
-                title="Supprimer"
-              >
-                <Trash2 className="h-4 w-4" />
-              </button>
-            }
-            pseudo={user.profile?.pseudo}
-            onConfirm={() => onDelete(user.id)}
-            onCancel={() => {}}
-          />
-        </div>
+        <UserActions user={user} currentUser={currentUser} refresh={refresh} />
       );
     },
   },
